# fix: delete advisory performance